Extract _cli.py + _templates.py from _legacy.py (#1404)
